+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Message Translator Change Log
 
-## 2.1.1 under development
+## 2.1.1 November 23, 2022
 
 - Bug #93: Throw exceptions on empty and not passed parameters, cast `null` to empty string in `SimpleMessageFormatter` 
   (@arogachev)